# vulkan: optimization proposals for coopmat1 mul_mm