@0cc4m 2025-03-11
> https://github.com/ggml-org/llama.cpp/commit/da0d6989de4229b7388f486e25a37cc5687919a1 : use f16 for shared buffers if possible (actually no impact on performance, but can reduce shared memory consumption)

This is already the case, coopmat requires fp16 and that means the buffers will always be 16-bit floats. I think the commit doesn't change anything.