@0cc4m 2025-05-10
@remyoudompheng Are you still around? Optimizing coopmat1 is a good idea, let's get this PR ready.